refactor(sync): per-topic split of claude adapter trait methods
FU-2 follow-up to T3.3. Pulls the 6 remaining trait method bodies
out of claude/mod.rs into topical submodules following the
plugin_assets.rs / skills.rs pattern. Each trait method becomes
a single-line delegation; the inherent impl block keeps only the
constructor and dir-helper accessors.

New submodules (each pub(super) free fns named *_impl):
- claude/commands.rs     (read_commands, write_commands, collect_commands_from_dir helper)
- claude/settings.rs     (read/write mcp_servers + read/write preferences — share settings.json)
- claude/hooks.rs        (read/write hooks)
- claude/agents.rs       (read/write agents)
- claude/instructions.rs (read/write instructions)

claude/mod.rs:    1631 -> 967 LOC (-41% in this commit, -53% from
                  the original 2045 LOC file before T3.3)
8-file split: 967 + 7 submodules (152-253 LOC each), each
focused on one concern with its own module docstring.

312 sync-lib tests pass. Test-only imports (`SystemTime`,
`ContentFormat`, `hash_content`, `McpTransport`) are
`#[cfg(test)]`-gated since they are only referenced by tests
via `use super::*`.
